requirements.txt
selenium>=4.18.1
python-dotenv>=0.19.0
webdriver-manager==4.0.1 